Send along `is_test=1` when configured for test.
This is a new (or at least undocumented) feature of Akismet that tells
Akismet __not__ to learn from this request. This is useful for testing
libraries or Akismet integrations. Per a discussion with Alex (an
Akismet engineer), they will be blogging about this feature in the next
couple of days, but it is already supported.