Move :permalink from default mappings to request mappings
Commit 81460d2beb94 added the sending of permalink to Akismet by adding
it to the list of "default mappings" all of which get `comment_`
prepended. However, according to the Askismet documentation[1], it
appears that they expect `permalink` to be simply `permalink` and not
`comment_permalink`.

This commit fixes this discrepancy by moving the `:permalink` symbol
from the list of attributes that receives the prepended `comment_` to
the list of attributes that does not receive this prependage.

[1]: http://akismet.com/development/api/#comment-check